fix: achieve clean compile on macOS
Fix C++20 compile errors across the codebase:
- BitStream.h: remove invalid rvalue-to-lvalue casts
- Platform.h: add aarch64/ARM to little-endian detection
- dpxslice.h, gfcStructures.h: remove deprecated 'register' keyword
- DS_BinarySearchTree.h: add this-> for dependent base class members
- PacketLogger.cpp: add space between string literal and macro
- RakPeer.cpp: use unsigned int for hex constants
- gfcPlate.cpp: cast GLhandleARB to GLuint, fix GL_CLAMP_TO_EDGE_EXT
- gfcSequence.cpp, gfcTrack.cpp: use unique_lock for condition_variable::wait
- main.cpp: fix intptr_t cast for GLhandleARB comparison
- playlistwindow.cpp: fix narrowing conversion
- gfcimagesaver.cpp: stub out removed GFL/EXR savers
- gfcpixelbuffer.h: add GFLC_BITMAP stub methods for trilerp.cpp
- CMakeLists.txt: fix platform exclusions, add Imath, exclude duplicate gfcTrack.cpp
- gfcimageloaderexr: exclude from build (OIIO handles EXR)
- Regenerate GLAD for compatibility profile (legacy OpenGL support)

+++ src/main.cpp
@@ -561,7 +561,7 @@ int main(int argc, char *argv[]) {
         int waitingForTestProgramCounter=0;
         int testProgramCounterLimit=50;
         //printf(" Test Program Object=%i\n",testProgramObject);
-        while ((int)testProgramObject==0 && waitingForTestProgramCounter<=testProgramCounterLimit) {
+        while ((intptr_t)testProgramObject==0 && waitingForTestProgramCounter<=testProgramCounterLimit) {
             //printf("counter=%i \n",waitingForTestProgramCounter);
             mw.mainWindow->redraw();
             //printf("main window redrawn \n");
